1- refactor Login 2- implement the state and cubit 3- add data sourc
