+++ lib/feature/auth/forget_password/data/data_source/forget_password_data_source.dart
@@ -1,17 +1,17 @@
 import 'package:dartz/dartz.dart';
-import 'package:telegram/core/di/service_locator.dart';
 import 'package:telegram/core/error/faliure.dart';
 import 'package:telegram/core/network/api/api_service.dart';
 
-abstract class ForgetPasswordRemoteDataSource {
+abstract class ForgetPasswordDataSource {
   Future<Either<Failure, void>> forgetPassword(String email);
   Future<Either<Failure, void>> resetPassword(String token, String newPassword);
+  Future<Either<Failure, void>> logoutFromAllDevices();
 }
 
-class ForgetPasswordRemoteDataSourceImpl
-    implements ForgetPasswordRemoteDataSource {
-  ForgetPasswordRemoteDataSourceImpl();
-  final ApiService apiService = sl<ApiService>();
+class ForgetPasswordDataSourceImp
+    implements ForgetPasswordDataSource {
+  ForgetPasswordDataSourceImp(this.apiService);
+  final ApiService apiService ;
 
   @override
   Future<Either<Failure, void>> forgetPassword(String email) async {
@@ -36,6 +36,21 @@ class ForgetPasswordRemoteDataSourceImpl
       final response = await apiService.post(
         endPoint: '',
         data: {'token': token, 'newPassword': newPassword},
+      );
+      // Assuming a successful response returns an empty right value
+      return Right(null);
+    } catch (error) {
+      // Handle error and return a Failure
+      return Left(error as Failure);
+    }
+  }
+
+  @override
+  Future<Either<Failure, void>> logoutFromAllDevices() async {
+    try {
+      final response = await apiService.post(
+        endPoint: '',
+        data: {},
       );
       // Assuming a successful response returns an empty right value
       return Right(null);
