+++ lib/feature/auth/login/domain/repositories/base_repo.dart
@@ -3,6 +3,6 @@ import 'package:telegram/feature/auth/login/data/model/login_request_model.dart'
 
 import '../../../../../core/error/faliure.dart';
 
-abstract class BaseLoginRepository {
+abstract class LoginRepository {
   Future<Either<Failure, Unit>> login({required LoginRequestBody loginModel});
 }
